# Conflicts:
#	identityprovider/src/main/java/nz/ac/canterbury/seng302/identityprovider/database/UserModel.java
#	identityprovider/src/main/java/nz/ac/canterbury/seng302/identityprovider/database/UserRepository.java
#	identityprovider/src/main/java/nz/ac/canterbury/seng302/identityprovider/service/RegisterServerService.java
#	portfolio/src/main/java/nz/ac/canterbury/seng302/portfolio/controller/ViewAccountController.java
#	portfolio/src/main/java/nz/ac/canterbury/seng302/portfolio/model/contract/UserContract.java
#	portfolio/src/main/java/nz/ac/canterbury/seng302/portfolio/service/UserAccountService.java
#	portfolio/src/test/java/nz/ac/canterbury/seng302/portfolio/controller/ViewAccountControllerTest.java
#	portfolio/subdirectory/demodb.mv.db
#	portfolio/subdirectory/demodb.trace.db

+++ identityprovider/src/main/java/nz/ac/canterbury/seng302/identityprovider/database/UserRepository.java
@@ -1,11 +1,10 @@
 package nz.ac.canterbury.seng302.identityprovider.database;
 
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserRepository extends CrudRepository<UserModel, Long> {
+public interface UserRepository extends CrudRepository<UserModel, Integer> {
 
-  UserModel findByUsername(String username);
+    UserModel findByUsername(String username);
 
-  UserModel findById(int id);
+    UserModel findById(int id);
 }
